Recompiled the app with the latest version of the add-on sdk
This add-on had stopped working with v38 and on.  I just now found time
to recompile it using the latest version of jQuery and got rid of the
iframe layover that was dependent on mouse position.  I replaced that
with a fixed CSS div wrapper.  I also updated the search URL buttons
with https for google scholar and bing because they used to not use
secure and added yahoo to the list.